+++ main/actuators/peltier.h
@@ -21,7 +21,7 @@ void PeltierInit(digitalPin csPin);
 void PeltierSet(digitalPin csPin, UINT8 lShdn, int set);
 
 
-int PeltierPID(float ref, float y);
+float PeltierPID(float ref, float y);
 
 void PeltierResetPID(void);
